+++ src/plugin/systems.rs
@@ -25,7 +25,7 @@ use rapier::prelude::*;
 use std::collections::HashMap;
 
 #[cfg(feature = "dim3")]
-use crate::prelude::AsyncCollider;
+use crate::prelude::{AsyncCollider, AsyncSceneCollider};
 
 /// Components that will be updated after a physics step.
 pub type RigidBodyWritebackComponents<'a> = (
@@ -537,31 +537,77 @@ pub fn step_simulation<PhysicsHooksData: 'static + WorldQuery + Send + Sync>(
 
 /// NOTE: This currently does nothing in 2D.
 #[cfg(feature = "dim2")]
-pub fn init_async_shapes() {}
+pub fn init_async_colliders() {}
 
 /// System responsible for creating `Collider` components from `AsyncCollider` components if the
-/// corresponding resource has become available.
+/// corresponding mesh has become available.
 #[cfg(feature = "dim3")]
-pub fn init_async_shapes(
+pub fn init_async_colliders(
     mut commands: Commands,
     meshes: Res<Assets<Mesh>>,
-    async_shapes: Query<(Entity, &AsyncCollider), Without<Collider>>,
+    async_colliders: Query<(Entity, &AsyncCollider)>,
 ) {
-    for (entity, async_shape) in async_shapes.iter() {
-        match async_shape {
-            AsyncCollider::Mesh(handle) => {
-                if let Some(coll) = meshes.get(handle).and_then(Collider::bevy_mesh) {
-                    commands.entity(entity).insert(coll);
+    for (entity, async_collider) in async_colliders.iter() {
+        if let Some(mesh) = meshes.get(&async_collider.handle) {
+            match Collider::from_bevy_mesh(mesh, &async_collider.shape) {
+                Some(collider) => {
+                    commands
+                        .entity(entity)
+                        .insert(collider)
+                        .remove::<AsyncCollider>();
                 }
+                None => error!("Unable to generate collider from mesh {:?}", mesh),
             }
-            AsyncCollider::ConvexDecomposition(handle, params) => {
-                if let Some(coll) = meshes
-                    .get(handle)
-                    .and_then(|m| Collider::bevy_mesh_convex_decomposition_with_params(m, params))
-                {
-                    commands.entity(entity).insert(coll);
+        }
+    }
+}
+
+/// System responsible for creating `Collider` components from `AsyncSceneCollider` components if the
+/// corresponding scene has become available.
+#[cfg(feature = "dim3")]
+pub fn init_async_scene_colliders(
+    mut commands: Commands,
+    meshes: Res<Assets<Mesh>>,
+    scenes: Res<Assets<Scene>>,
+    async_colliders: Query<(Entity, &AsyncSceneCollider)>,
+    children: Query<&Children>,
+    mesh_handles: Query<(&Name, &Handle<Mesh>)>,
+) {
+    for (entity, async_collider) in async_colliders.iter() {
+        if scenes.get(&async_collider.handle).is_some() {
+            traverse_descendants(entity, &children, &mut |child| {
+                if let Ok((name, handle)) = mesh_handles.get(child) {
+                    let shape = async_collider
+                        .named_shapes
+                        .get(name.as_str())
+                        .unwrap_or(&async_collider.shape);
+                    if let Some(shape) = shape {
+                        let mesh = meshes.get(handle).unwrap(); // NOTE: Mesh is already loaded
+                        match Collider::from_bevy_mesh(mesh, shape) {
+                            Some(collider) => {
+                                commands.entity(child).insert(collider);
+                            }
+                            None => error!(
+                                "Unable to generate collider from mesh {:?} with name {}",
+                                mesh, name
+                            ),
+                        }
+                    }
                 }
-            }
+            });
+
+            commands.entity(entity).remove::<AsyncSceneCollider>();
+        }
+    }
+}
+
+/// Iterates over all descendants of the `entity` and applies `f`.
+#[cfg(feature = "dim3")]
+fn traverse_descendants(entity: Entity, children: &Query<&Children>, f: &mut impl FnMut(Entity)) {
+    if let Ok(entity_children) = children.get(entity) {
+        for child in entity_children.iter().copied() {
+            f(child);
+            traverse_descendants(child, children, f);
         }
     }
 }
@@ -998,9 +1044,20 @@ pub fn update_colliding_entities(
 
 #[cfg(test)]
 mod tests {
-    use bevy::ecs::event::Events;
+    #[cfg(feature = "dim3")]
+    use bevy::prelude::shape::{Capsule, Cube};
+    use bevy::{
+        asset::AssetPlugin,
+        core::CorePlugin,
+        ecs::event::Events,
+        render::{settings::WgpuSettings, RenderPlugin},
+        scene::ScenePlugin,
+        window::WindowPlugin,
+    };
 
     use super::*;
+    #[cfg(feature = "dim3")]
+    use crate::prelude::ComputedColliderShape;
 
     #[test]
     fn colliding_entities_updates() {
@@ -1086,5 +1143,109 @@ mod tests {
             colliding_entities2.is_empty(),
             "Colliding entity should be removed from the CollidingEntities component when the collision ends"
         );
+    }
+
+    #[test]
+    #[cfg(feature = "dim3")]
+    fn async_collider_initializes() {
+        let mut app = App::new();
+        app.add_plugin(HeadlessRenderPlugin)
+            .add_system(init_async_colliders);
+
+        let mut meshes = app.world.resource_mut::<Assets<Mesh>>();
+        let cube = meshes.add(Cube::default().into());
+
+        let entity = app
+            .world
+            .spawn()
+            .insert(AsyncCollider {
+                handle: cube,
+                shape: ComputedColliderShape::TriMesh,
+            })
+            .id();
+
+        app.update();
+
+        let entity = app.world.entity(entity);
+        assert!(
+            entity.get::<Collider>().is_some(),
+            "Collider component should be added"
+        );
+        assert!(
+            entity.get::<AsyncCollider>().is_none(),
+            "AsyncCollider component should be removed after Collider component creation"
+        );
+    }
+
+    #[test]
+    #[cfg(feature = "dim3")]
+    fn async_scene_collider_initializes() {
+        let mut app = App::new();
+        app.add_plugin(HeadlessRenderPlugin)
+            .add_system(init_async_scene_colliders);
+
+        let mut meshes = app.world.resource_mut::<Assets<Mesh>>();
+        let cube_handle = meshes.add(Cube::default().into());
+        let capsule_handle = meshes.add(Capsule::default().into());
+        let cube = app
+            .world
+            .spawn()
+            .insert(Name::new("Cube"))
+            .insert(cube_handle)
+            .id();
+        let capsule = app
+            .world
+            .spawn()
+            .insert(Name::new("Capsule"))
+            .insert(capsule_handle)
+            .id();
+
+        let mut scenes = app.world.resource_mut::<Assets<Scene>>();
+        let scene = scenes.add(Scene::new(World::new()));
+
+        let mut named_shapes = bevy::utils::HashMap::new();
+        named_shapes.insert("Capsule".to_string(), None);
+        let parent = app
+            .world
+            .spawn()
+            .insert(AsyncSceneCollider {
+                handle: scene,
+                shape: Some(ComputedColliderShape::TriMesh),
+                named_shapes,
+            })
+            .push_children(&[cube, capsule])
+            .id();
+
+        app.update();
+
+        assert!(
+            app.world.entity(cube).get::<Collider>().is_some(),
+            "Collider component should be added for cube"
+        );
+        assert!(
+            app.world.entity(capsule).get::<Collider>().is_none(),
+            "Collider component shouldn't be added for capsule"
+        );
+        assert!(
+            app.world.entity(parent).get::<AsyncCollider>().is_none(),
+            "AsyncSceneCollider component should be removed after Collider components creation"
+        );
+    }
+
+    // Allows run tests for systems containing rendering related things without GPU
+    struct HeadlessRenderPlugin;
+
+    impl Plugin for HeadlessRenderPlugin {
+        fn build(&self, app: &mut App) {
+            app.insert_resource(WgpuSettings {
+                backends: None,
+                ..WgpuSettings::default()
+            })
+            .add_plugin(CorePlugin::default())
+            .add_plugin(WindowPlugin::default())
+            .add_plugin(AssetPlugin::default())
+            .add_plugin(ScenePlugin::default())
+            .add_plugin(RenderPlugin::default());
+        }
     }
 }